requirements.txt
# Core
fastapi==0.95.2
uvicorn==0.21.1
python-multipart==0.0.6

# Image Processing
opencv-python-headless==4.8.1.78
numpy==1.26.4
pdf2image==1.16.3
imutils==0.5.4
Pillow==10.0.0

# Machine Learning
scikit-learn==1.3.0

# File Handling
aiofiles==23.2.1

# Data Validation
pydantic==1.10.13
python-dotenv==1.0.0